Changed namespace and moved VO
