+++ module/API/src/V1/Rpc/Login/LoginController.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
 
 namespace API\V1\Rpc\Login;
 
-use EyeChart\VO\AuthenticationVO;
+use EyeChart\VO\Authentication\AuthenticationVO;
 use EyeChart\VO\VOInterface;
 use Zend\Mvc\Controller\AbstractActionController;
 use EyeChart\Service\Authenticate\AuthenticateService;
